Break out VenueHistory component and lastFMApiCalls
