+++ src/components/ArtistShow/ArtistShow.js
@@ -1,125 +1,72 @@
 import React, { Component, Fragment } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import Nav from '../Nav/Nav';
-import './ArtistShow.css';
-import { getArtist } from '../../Utils/backendApiCalls';
-import {
-  getArtistDetails,
-  getEventHistory
-} from '../../Utils/songKickApiCalls';
 import { Link } from 'react-router-dom';
+
+import { getArtist } from '../../Utils/backendApiCalls';
+import { getLastFMInfo } from '../../Utils/lastFM/lastFMApiCalls';
+import './ArtistShow.css';
+
+import Nav from '../Nav/Nav';
 import LoadingSpinner from '../LoadingSpinner/LoadingSpinner';
+import VenueHistory from '../artistInfoCommon/VenueHistory/VenueHistory';
 
 export class ArtistShow extends Component {
   state = {
-    name: '',
-    image_url: '',
-    bio: '',
-    genres: [],
-    similarArtists: [],
-    spotify_url: '',
-    localEventHistory: {}
+    artist: {}
   };
 
   async componentDidMount() {
     const { id } = this.props;
     const artist = await getArtist(id);
-    const artistDetails = await getArtistDetails(artist.name);
-    const cleanArtist = await this.formatArtist(artistDetails);
-    const localEventHistory = await getEventHistory(
-      artist.songkick_id,
-      this.props.currentVenue,
-      artist.name
-    );
+    const lastFMInfo = await getLastFMInfo(artist.name);
 
-    await this.setState({
-      name: artist.name,
-      image_url: cleanArtist.image,
-      bio: cleanArtist.bio,
-      spotify_followers: artist.spotify_followers,
-      genres: cleanArtist.genres,
-      similarArtists: cleanArtist.similarArtists,
-      spotify_url: artist.spotify_url,
-      localEventHistory
+    this.setState({
+      artist: {
+        ...artist,
+        ...lastFMInfo
+      }
     });
   }
 
-  formatArtist = artist => {
-    const imageValues = Object.values(artist.artist.image[3]);
-    const image = imageValues[0];
-
-    const bioData = artist.artist.bio.summary.split('<');
-    const trimData = bioData[0].trim();
-    let bio = bioData[0];
-
-    if (trimData.length < 40) {
-      bio = 'Artist biography unavailable';
-    }
-
-    const genres = artist.artist.tags.tag.splice(0, 4);
-    const similarArtists = artist.artist.similar.artist.splice(0, 4);
-    const formattedArtist = {
-      bio,
-      image,
-      genres,
-      similarArtists
-    };
-    return formattedArtist;
-  };
-
   render() {
+    const { artist } = this.state;
+    const { currentVenue } = this.props;
     const {
       name,
-      image_url,
+      image,
       bio,
       spotify_followers,
       genres,
       similarArtists,
-      spotify_url,
-      localEventHistory
-    } = this.state;
+      spotify_url
+    } = this.state.artist;
 
-    const displayGenres = genres.map((genre, index) => (
-      <p className="genre" key={genre + index}>
-        {genre.name}
-      </p>
-    ));
-
-    const similarArtistDisplay = similarArtists.map((artist, index) => (
-      <p className="similar" key={`${artist.name}-${index}`}>
-        {artist.name}
-      </p>
-    ));
-
-    let displayHistory;
-
-    if (localEventHistory.events) {
-      if (localEventHistory.events.length) {
-        displayHistory = localEventHistory.events.map((event, i) => (
-          <div className="table-row" key={`${event.id}-${i}`}>
-            <h4 className="event-details">{event.date}</h4>
-            <h4 className="event-details"> {event.venue.displayName}</h4>
-            <h4 className="event-details"> {event.billing}</h4>
-          </div>
-        ));
-      }
-    }
-
+    let displayGenres;
+    let similarArtistDisplay;
     let displayArtist;
 
-    if (!localEventHistory.events || !similarArtists.length || !genres.length) {
+    if (!similarArtists || !genres) {
       displayArtist = <LoadingSpinner />;
+      displayGenres = <LoadingSpinner />;
+      similarArtistDisplay = <LoadingSpinner />;
     } else {
+      displayGenres = genres.map((genre, index) => (
+        <p className="genre" key={genre + index}>
+          {genre.name}
+        </p>
+      ));
+
+      similarArtistDisplay = similarArtists.map((artist, index) => (
+        <p className="similar" key={`${artist.name}-${index}`}>
+          {artist.name}
+        </p>
+      ));
       displayArtist = (
         <Fragment>
           <div className="portrait-wrap">
             <div className="profile-frame">
-              <img
-                className="profile-image"
-                src={image_url}
-                alt="artist-profile"
-              />
+              <img className="profile-image" src={image} alt="artist-profile" />
             </div>
             <div className="stat-wrap">
               <span className="name-wrap">
@@ -156,22 +103,7 @@ export class ArtistShow extends Component {
             <h3 className="similar-artists">Similar Artists:</h3>
             {similarArtistDisplay}
           </span>
-          <div className="past-events">
-            <div className="watchlist">
-              <h3>
-                Latest Appearances In
-                <span className="city-header">
-                  {localEventHistory.venueCity}
-                </span>
-              </h3>
-              <div className="recent-table-headings">
-                <h4 className="event-header">Date</h4>
-                <h4 className="event-header">Venue</h4>
-                <h4 className="event-header">Billing</h4>
-              </div>
-              <div className="event-container">{displayHistory}</div>
-            </div>
-          </div>
+          <VenueHistory artist={artist} currentVenue={currentVenue} />
         </Fragment>
       );
     }
